Cut suite peak memory 2.4x -> 1.6x; auto-downsize instead of skipping
Reviewing 6fdd697: the guard I added worked correctly -- it excluded the stale
1,500-user files from T2 rather than mixing scales, and T2 is honest. But all
three large corpora SKIPPED (needed 38 GB, 25 GB free), so the paper is back to
one corpus in its tables. The guard prevented a wrong result; it did not
produce a right one.

Two real memory wins:
  - neural baselines called full_catalog_metrics THREE times per user while
    holding the full u x i score matrix; now one pass, then immediate release
    (~6.4 GB on yelp2018)
  - the per-user feature blocks were float64; float32 halves the largest
    resident structure. The Gram accumulation promotes to float64, so the
    fitted weights are unaffected -- verified efficiency 8.7e-19 and identical
    Shapley values.

Peak drops from ~2.4x to ~1.6x the score-matrix footprint.

Notebook now AUTO-DOWNSIZES rather than skipping: at 25 GB free it runs
yelp2018 at 16,426 users, gowalla at 15,250, amazon_book at 7,222 -- roughly
10x the stale sandbox files. A smaller real run beats no run, and beats
leaving a stale file that later contaminates the tables.

+++ src/signalshap/game/core.py
@@ -111,9 +111,12 @@ class SignalShapGame:
         self.feat: dict[int, np.ndarray] = {}
         for u in set(self.eval_users) | set(self.fit_users):
             c = candidates[u]
+            # float32 halves the largest resident structure; the ridge solve
+            # promotes to float64 via the Gram accumulation, so precision of
+            # the fitted weights is unaffected.
             self.feat[u] = np.column_stack(
                 [z_normalise(scores[g][u][c]) for g in self.sources]
-            ).astype(np.float64)
+            ).astype(np.float32)
 
         self._v0 = self._frozen_baseline(v0_seed)
         self._cache: dict[frozenset, float] = {}
@@ -158,7 +161,7 @@ class SignalShapGame:
             XtX = np.zeros((n, n))
             Xty = np.zeros(n)
             for u in self.fit_users:
-                F = self.feat[u]
+                F = self.feat[u].astype(np.float64, copy=False)
                 y = (self.candidates[u] == self.valid_items[u]).astype(np.float64)
                 XtX += F.T @ F
                 Xty += F.T @ y
